+++ karma.conf.js
@@ -115,7 +115,7 @@ module.exports = function(config) {
 
     // start these browsers
     // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-    //browsers: ['PhantomJS', 'Chrome', 'Firefox', 'Safari'],
+    //browsers: ['ChromeHeadless', 'Chrome', 'Firefox', 'Safari'],
     browsers: ['ChromeHeadless'],
 
     customLaunchers: {
@@ -127,12 +127,6 @@ module.exports = function(config) {
         base: 'IE',
         'x-ua-compatible': 'IE=EmulateIE8'
       }
-    },
-
-    phantomjsLauncher: {
-      // Have phantomjs exit if a ResourceError is encountered (useful if karma
-      // exits without killing phantom)
-      exitOnResourceError: true
     },
 
     // Continuous Integration mode